requirements.txt
numpy>=1.19.0
pyzmq>=22.0.0
matplotlib>=3.3.0
psutil>=5.8.0
